Cart to Bookings transfer on pay

+++ routes/cart.js
@@ -1,6 +1,7 @@
 import express from 'express';
 import Item from '../db/models/Item.js';
 import Trip from '../db/models/Trip.js';
+import Booking from '../db/models/Booking.js';
 import { validateTripIdInReqBody } from '../lib/helpers.js';
 const router = express.Router();
 
@@ -26,7 +27,7 @@ router.post('/addtrip', async (req, res) => {
       return;
     }
     if (result) res.json({ result: true, message: 'Trip saved' });
-  } else res.json({ result: false, error: 'Invalid trip data' });
+  } else res.json({ result: false, error: 'Invalid trip ID' });
 });
 
 router.get('/trips', async (req, res) => {
@@ -42,17 +43,52 @@ router.get('/trips', async (req, res) => {
     : res.json({ result: false, error: 'Failed to get trips' });
 });
 
-router.delete('/removetrip', async (req, res) => {
+router.post('/paytrips', async (req, res) => {
   let status;
   try {
-    status = await Item.deleteOne({ trip_id: req.query.trip_id });
+    status = await Item.updateMany({}, { payed: true });
+    console.log(status);
   } catch (err) {
     console.log(err);
-    return res.json({ result: false, error: 'Failed to delete trip' });
+    return res.json({ result: false, error: 'Failed to update pay status' });
   }
-  status.deletedCount > 0
-    ? res.json({ result: true, message: 'Trip deleted' })
-    : res.json({ result: false, error: 'Trip not found' });
+  const payedItems = await Item.find();
+  if (status.modifiedCount === payedItems.length || status.matchedCount === payedItems.length) {
+    try {
+      for (let { trip_id, payed } of payedItems) {
+        let foundTrip;
+        try {
+          foundTrip = await Booking.findOne({ trip_id });
+        } catch (err) {
+          console.log(err);
+          res.json({ result: false, error: 'Error with trip comparison' });
+          return;
+        }
+        if (foundTrip) return res.json({ result: false, error: 'Trip already stored in bookings' });
+        await new Booking({ trip_id, payed }).save();
+      }
+      await Item.deleteMany({});
+    } catch (err) {
+      console.log(err);
+      res.json({ result: false, error: 'Failed to transfer all payed trips to bookings' });
+    }
+    res.json({ result: true, message: 'All payed trips transfered to bookings' });
+  } else res.json({ result: false, error: 'Trips not found' });
+});
+
+router.delete('/removetrip', async (req, res) => {
+  if (validateTripIdInReqBody(req.query.trip_id)) {
+    let status;
+    try {
+      status = await Item.deleteOne({ trip_id: req.query.trip_id });
+    } catch (err) {
+      console.log(err);
+      return res.json({ result: false, error: 'Failed to delete trip' });
+    }
+    status.deletedCount > 0
+      ? res.json({ result: true, message: 'Trip deleted' })
+      : res.json({ result: false, error: 'Trip not found' });
+  } else res.json({ result: false, error: 'Invalid trip ID' });
 });
 
 export default router;
